+++ internal/pkg/service/common/goaextension/dependencies/dependencies.go
