+++ internal/pkg/service/common/goaextension/genericerror/error.go
