+++ internal/pkg/service/common/goaextension/oneof/oneof.go
