+++ internal/pkg/service/common/goaextension/operationid/operationid.go
