+++ internal/pkg/service/common/goaextension/operationid/operationid_test.go
